# perf: delay initial render of PopoverComponents in SearchFiltersBar